Refactor invitation handling and UI improvements
The existing code for handling invitations was refactored for clarity
and efficiency. This included enhancing the functions responsible for
patching invitations into database events and adding conditions for
handling default cases. We streamlined the process, reducing code
redundancies and removing unsupported cases within the WorkOS
typescript types.

In the user interface, we simplified and improved how user actions
are executed and notified via the toast messages. This ensures
more accurate updates regarding the success or failure of actions like
resending or revoking invitations.

Additionally, a new component 'OnlineUsersIndicator' was introduced to
display the number of online users. This enhancement provides
dynamic user engagement updates in real-time, improving the user
experience and operational transparency for organization members.